The FAST library compiles but the C++ driver program doesn't

+++ fastv8/Source/dependencies/OpenFOAM/OpenFOAM_Types.h
@@ -54,8 +54,8 @@
   typedef struct OpFM_ParameterType {
     void * object ;
     float AirDens ;
-    float NumBl ;
-    float NMappings ;
+    int NumBl ;
+    int NMappings ;
     int NnodesVel ;
     int NnodesForce ;
     int NnodesForceBlade ;
